requirements.txt
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
Django==3.2.2
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
